+++ src/dom/attribute.ts
@@ -6,7 +6,11 @@ export class Attribute {
 	public readonly value: string | DynamicValue;
 	public readonly location: Location;
 
-	constructor(key: string, value: string | DynamicValue, location: Location) {
+	public constructor(
+		key: string,
+		value: string | DynamicValue,
+		location: Location
+	) {
 		this.key = key;
 		this.value = value;
 		this.location = location;
@@ -14,6 +18,29 @@ export class Attribute {
 		/* force undefined to null */
 		if (typeof this.value === "undefined") {
 			this.value = null;
+		}
+	}
+
+	/**
+	 * Test attribute value.
+	 *
+	 * @param {RegExp|string} pattern - Pattern to match value against. RegExp or
+	 * a string (===)
+	 * @returns {boolean} `true` if attribute value matches pattern.
+	 */
+	public valueMatches(pattern: RegExp): boolean;
+	public valueMatches(pattern: string): boolean;
+	public valueMatches(pattern: any): boolean {
+		/* dynamic values matches everything */
+		if (this.value instanceof DynamicValue) {
+			return true;
+		}
+
+		/* test value against pattern */
+		if (pattern instanceof RegExp) {
+			return this.value.match(pattern) !== null;
+		} else {
+			return this.value === pattern;
 		}
 	}
 }